Watch all the files in the directory.
We now store a class veriable called `_times` which is a dict mapping
path -> time. Instead of just checking the one file path, we check
everything in this dict. We build the dict of files to watch from the
function `get_paths` which has been modified to output files as well as
dicts.