Initial restructure for modifing output.
What was the `reload_script` function became `_reload_script_module`, and
what was `reload-with-py` is now `reload_script`. The new `reload_script`
function is called every time now(instead of just whenever `Use py console`
is checked) and the output is always tracked. There is also a new class
`SplitIO` which is a subclass of `StringIO` but it modifies the input
stream and it sends a stream to sys.stdout(or sys.stderr, or whatever
stream you give the constructor).