feat: proof-of-concept thread ping-pong for Pdb
Pdb is run inside it's own thread, but never concurrently with the main
thread. Blender is doesn't have very good support for threading in
python[1] so it would cause errors to run Pdb in it's own thread
unchecked. If it wasn't run in a thread it would lock up blender.

To get around both of these issues, Pdb is run inside a thread ping-pong
system that ensures that only it's thread or the main thread are active
at any given moment. When Pdb encounters a breakpoint it's thread will
pause and wait to be resumed. When the user gives the next command it
will be passed to Pdb, the main thread will be paused and the Pdb thread
will resume.

The flow of events is like this:

1. Main thread spawns Pdb thread and starts the debugger
2. Main thread pauses and waits for the pdb thread
3. When Pdb hits a breakpoint it will pause and resume the main thread
4. When the user gives user input, the input will be passed to the Pdb
thread
5. Main thread pauses and resumes the Pdb thread
6. Repeat steps 3-5

[1]: https://docs.blender.org/api/blender_python_api_current/info_gotcha.html#strange-errors-using-threading-module